+++ inc/x86_64/space_dma.hpp
@@ -60,5 +60,5 @@ class Space_dma final : public Space_mem<Space_dma>
 
         auto get_sdid() const { return sdid; }
 
-        static void access_ctrl (uint64_t phys, size_t size, Paging::Permissions perm) { Space_mem::access_ctrl (nova, phys, size, perm, Memattr::ram()); }
+        static void access_ctrl (uintptr_t phys, size_t size, Paging::Permissions perm) { Space_mem::access_ctrl (nova, phys, size, perm, Memattr::ram()); }
 };
